+++ frontend/src/stylesheet/components/_sidebar.scss
@@ -6,6 +6,7 @@
   height: 100%;
   display: flex;
   align-items: stretch;
+  overflow-y: auto;
 }
 
 .sidebar-main * {
